+++ webcomponents/markdown/CHANGELOG.md
@@ -1,5 +1,9 @@
 # 1.1.0 (2020-11-15)
 
+### Features
+
+- add a method to expose the container
+
 ### Style
 
 - improve font styling